:sparkles: Create TouchChecker class for further development
