+++ lib/src/pose_processor.dart
@@ -1,6 +1,7 @@
-import 'package:fca_pose_validation/src/angle_calculator.dart';
 import 'package:google_ml_kit/google_ml_kit.dart';
-import 'exercise_controller.dart';
+import 'exercise_definition.dart';
+import 'angle_calculator.dart';
+import 'touch_checker.dart';
 
 enum PoseProcessorType{
   angle,
@@ -19,8 +20,10 @@ class PoseProcessorError{
 
 class PoseProcessor {
   AngleCalculator angle = AngleCalculator();
+  TouchChecker touch = TouchChecker();
 
   void setPose(Pose pose){
     angle.setPose(pose);
+    touch.setPose(pose);
   }
 }
