+++ lib/src/touch_checker.dart
@@ -1,0 +1,11 @@
+import 'package:google_ml_kit/google_ml_kit.dart';
+
+class TouchChecker{
+  late Pose _pose;
+
+  void setPose(Pose pose){
+    _pose = pose;
+  }
+
+
+}
